save before r fix

+++ web/core/static/core/store/remake.js
@@ -8,33 +8,37 @@ import {make_id, random_vector, theme} from '../app.js';
 
 export const create_remake_slice = (set,get)=>({remake:{
     copy(d, src, a){ // maybe place in d.node (only run for part
-        const cpy = d.make.node(d, d.n[src].m, d.n[src].t);
-        if(d.n[src].m != 'p') d.n[cpy].v = d.n[src].v;
-        d.node.for_r(d, src, r=>{ // make for_rn that just has these nested loops
-            if(!(a&&a.exclude_r && a.exclude_r == r) && !(a&&a.r_t && !a.r_t.includes(d.n[r].t))){
-                d.node.for_n(d, r, (r,n,t,o)=>{
-                    if(src == n) d.make.edge(d, r, cpy, {t:t, o:o}); // adding edge in edge loop bad?!?!?!
-                });
-            }
-        });
-        d.node.for_n(d, src, (r,n,t,o)=>{
-            if(a&&a.deep) {
-                d.make.edge(d, cpy, d.remake.copy(d,n, {deep:true, exclude_r:r}), {t:t, o:o});
-            }else{
-                d.make.edge(d, cpy, n, {t:t, o:o});
-            }
-        });
-        d.next('reckon.node', cpy); // maybe this should go in node creation
-        return cpy;
+        if(!d.node.limited(d, [src])){
+            const cpy = d.make.node(d, d.n[src].m, d.n[src].t);
+            if(d.n[src].m != 'p') d.n[cpy].v = d.n[src].v;
+            d.node.for_r(d, src, r=>{ // make for_rn that just has these nested loops
+                if(!(a&&a.exclude_r && a.exclude_r == r) && !(a&&a.rt && !a.rt.includes(d.n[r].t))){
+                    d.node.for_n(d, r, (r,n,t,o)=>{
+                        if(src == n) d.make.edge(d, r, cpy, {t:t, o:o}); // adding edge in edge loop bad?!?!?!
+                    });
+                }
+            });
+            d.node.for_n(d, src, (r,n,t,o)=>{
+                if(a&&a.deep) {
+                    d.make.edge(d, cpy, d.remake.copy(d,n, {deep:true, exclude_r:r}), {t:t, o:o});
+                }else{
+                    d.make.edge(d, cpy, n, {t:t, o:o});
+                }
+            });
+            d.next('reckon.node', cpy); // maybe this should go in node creation
+            return cpy;
+        }
     },
     merge(d, nodes, target){ 
-        if(d.remake.merge_funcs[d.n[d.pick.same[0]].t]){  
-            d.remake.merge_funcs[d.n[d.pick.same[0]].t](d, nodes, target); 
-            d.remake.merge_funcs.base(d, nodes, target);  
-        }
-        else{  
-            d.remake.merge_funcs.default(d, nodes, target);  
-            d.remake.merge_funcs.base(d, nodes, target);
+        if(!d.node.limited(d, [...nodes, target])){
+            if(d.remake.merge_funcs[d.n[d.pick.same[0]].t]){  
+                d.remake.merge_funcs[d.n[d.pick.same[0]].t](d, nodes, target); 
+                d.remake.merge_funcs.base(d, nodes, target);  
+            }
+            else{  
+                d.remake.merge_funcs.default(d, nodes, target);  
+                d.remake.merge_funcs.base(d, nodes, target);
+            }
         }
     },
     merge_funcs:{
@@ -61,15 +65,17 @@ export const create_remake_slice = (set,get)=>({remake:{
         },
     },
     split(d, roots, target){ // make unique copy for everything but asset and transform?
-        const dead_edges = [];
-        d.node.for_n(d, roots, (r,n,t,o)=>{
-            if(n == target){
-                const cpy = d.remake.copy(d, n, {deep:true, r_t:['asset']}); // should include matrix node?
-                d.make.edge(d, r, cpy, {t:t, o:o});
-                dead_edges.push({r:r, t:t, o:o+1}); // +1 because new edge is inersted just in front of old edge 
-            }
-        });
-        d.node.delete_edges(d, dead_edges);
+        if(!d.node.limited(d, [...roots, target])){
+            const dead_edges = [];
+            d.node.for_n(d, roots, (r,n,t,o)=>{
+                if(n == target){
+                    const cpy = d.remake.copy(d, n, {deep:true, rt:['asset', 'view']}); // should include matrix node?
+                    d.make.edge(d, r, cpy, {t:t, o:o});
+                    dead_edges.push({r:r, n:n, t:t, o:o+1}); // +1 because new edge is inersted just in front of old edge 
+                }
+            });
+            d.node.delete_edges(d, dead_edges);
+        }
     },
 }});
